Add files via upload
These are labs 1 through 4. I'll go into detail on what each lab does.
Lab 1: The numbers X and Y are found at locations x3100 and x3101
Lab 2: The numbers X and Y are found at locations x3120 and x3121
Lab 3: Write a program in LC-3 assembly language that keeps prompting for an integer in the range 0-6, and each time it outputs the corresponding name of the day. If a key other than ’0’ through ’6’ is pressed, the program exits.
Lab 4: Write a program in LC-3 assembly language that computes Fn, the n−th Fibonacci number. Find the largest Fn such that no overflow occurs, i.e. find n = N such that FN is the largest Fibonacci number to be correctly represented with 16 bits in two’s complement format.